Move create identity to a service
We can remove the pseudonym and pseudonym seed complexity because pseudonyms are changing all the time
