+++ app/src/main/AndroidManifest.xml
@@ -110,5 +110,9 @@
             android:name=".EventPushService"
             android:exported="false" >
         </service>
+        <service
+            android:name=".CreateIdentityService"
+            android:exported="false" >
+        </service>
     </application>
 </manifest>
