+++ app/src/main/java/io/islnd/android/islnd/messaging/MessageLayer.java
@@ -16,10 +16,6 @@ import java.security.Key;
 public class MessageLayer {
     private static final String TAG = MessageLayer.class.getSimpleName();
 
-    public static String getPseudonym(Context context, String seed) {
-        return Rest.getPseudonym(seed, Util.getApiKey(context));
-    }
-
     public static boolean addFriendFromEncodedIdentityString(Context context,
                                                              String encodedString) {
         Log.v(TAG, "adding friend from encoded string: " + encodedString);
