+++ app/src/main/java/io/islnd/android/islnd/messaging/RestInterface.java
@@ -4,7 +4,6 @@ import io.islnd.android.islnd.messaging.crypto.EncryptedData;
 import io.islnd.android.islnd.messaging.crypto.EncryptedEvent;
 import io.islnd.android.islnd.messaging.server.EventQuery;
 import io.islnd.android.islnd.messaging.server.EventQueryResponse;
-import io.islnd.android.islnd.messaging.server.PseudonymResponse;
 
 import java.util.List;
 
@@ -19,11 +18,6 @@ public interface RestInterface {
     @GET("/readers/{username}")
     Call<List<EncryptedData>> readers(
             @Path("username") String username,
-            @Query("apiKey") String apiKey);
-
-    @GET("/pseudonym/{seed}")
-    Call<PseudonymResponse> pseduonym(
-            @Path("seed") String seed,
             @Query("apiKey") String apiKey);
 
     @POST("/publicKey/{username}")
