--- app/src/main/java/io/islnd/android/islnd/messaging/server/PseudonymResponse.java
@@ -1,9 +1,0 @@
-package io.islnd.android.islnd.messaging.server;
-
-public class PseudonymResponse {
-    private String pseudonym;
-
-    public String getPseudonym() {
-        return pseudonym;
-    }
-}
