+++ app/src/main/res/values/strings.xml
@@ -43,7 +43,6 @@
     <string name="private_key">PRIVATE_KEY</string>
     <string name="public_key">PUBLIC_KEY</string>
     <string name="group_key">group_key</string>
-    <string name="pseudonym_seed">pseudonym_seed</string>
     <string name="alias">alias</string>
 
     <string name="view_friend_status">Not Allowed</string>
